feat: circuit common data

+++ plonky2-verifier/src/fri/structure.cairo
@@ -9,16 +9,16 @@ use plonky2_verifier::fields::utils::{sum_array, max_array, pow, shift_left};
 
 #[derive(Drop)]
 pub struct FriOracleInfo {
-    pub num_polys: u32,
+    pub num_polys: usize,
     pub blinding: bool
 }
 
 #[derive(Drop)]
 pub struct FriPolynomialInfo {
     /// Index into `FriInstanceInfo`'s `oracles` list.
-    pub oracle_index: u32,
+    pub oracle_index: usize,
     /// Index of the polynomial within the oracle.
-    pub polynomial_index: u32,
+    pub polynomial_index: usize,
 }
 
 /// A batch of openings at a particular point.
@@ -57,22 +57,22 @@ pub struct FriChallenges {
     pub fri_betas: Array<GoldilocksQuadratic>,
     pub fri_pow_response: Goldilocks,
     // Indices at which the oracle is queried in FRI.
-    pub fri_oracle_indices: Array<u32>,
+    pub fri_oracle_indices: Array<usize>,
 }
 
-#[derive(Drop)]
+#[derive(Drop, Clone)]
 pub enum FriReductionStrategy {
-    Fixed: Array<u32>,
-    ConstantArityBits: (u32, u32),
+    Fixed: Array<usize>,
+    ConstantArityBits: (usize, usize),
 }
 
-#[derive(Drop)]
+#[derive(Drop, Clone)]
 pub struct FriConfig {
-    pub rate_bits: u32,
-    pub cap_height: u32,
-    pub proof_of_work_bits: u32,
+    pub rate_bits: usize,
+    pub cap_height: usize,
+    pub proof_of_work_bits: usize,
     pub reduction_strategy: FriReductionStrategy,
-    pub num_query_rounds: u32,
+    pub num_query_rounds: usize,
 }
 
 #[derive(Drop)]
@@ -81,12 +81,12 @@ pub struct FriParams {
     /// Whether to use a hiding variant of Merkle trees (where random salts are added to leaves).
     pub hiding: bool,
     /// The degree of the purported codeword, measured in bits.
-    pub degree_bits: u32,
+    pub degree_bits: usize,
     /// The arity of each FRI reduction step, expressed as the log2 of the actual arity.
     /// For example, `[3, 2, 1]` would describe a FRI reduction tree with 8-to-1 reduction, then
     /// a 4-to-1 reduction, then a 2-to-1 reduction. After these reductions, the reduced polynomial
     /// is sent directly.
-    pub reduction_arity_bits: Array<u32>,
+    pub reduction_arity_bits: Array<usize>,
 }
 
 
@@ -126,27 +126,27 @@ impl ReducingFactorImpl of ReducingFactorTrait {
 
 #[generate_trait]
 pub impl FriParamsImpl of FriParamsTrait {
-    fn total_arities(self: @FriParams) -> u32 {
+    fn total_arities(self: @FriParams) -> usize {
         sum_array(self.reduction_arity_bits.span())
     }
 
-    fn max_arity_bits(self: @FriParams) -> u32 {
+    fn max_arity_bits(self: @FriParams) -> usize {
         max_array(self.reduction_arity_bits.span())
     }
 
-    fn lde_bits(self: @FriParams) -> u32 {
+    fn lde_bits(self: @FriParams) -> usize {
         *self.degree_bits + *self.config.rate_bits
     }
 
-    fn lde_size(self: @FriParams) -> u32 {
+    fn lde_size(self: @FriParams) -> usize {
         shift_left(1, self.lde_bits())
     }
 
-    fn final_poly_bits(self: @FriParams) -> u32 {
+    fn final_poly_bits(self: @FriParams) -> usize {
         *self.degree_bits - self.total_arities()
     }
 
-    fn final_poly_len(self: @FriParams) -> u32 {
+    fn final_poly_len(self: @FriParams) -> usize {
         shift_left(1, self.final_poly_bits())
     }
 }
